myproject/requirements.txt
asgiref==3.8.1
Django
django-allauth==64.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
sqlparse==0.5.1
typing_extensions==4.12.2
dj-database-url==1.0.0
requests==2.31.0
psycopg2-binary==2.9.8  
cryptography
python-decouple
django-allauth
reportlab
stripe
